Flag md5 use as not used for security
beaker is computing the machine hash at import time. This usage of md5
is not flagged as not used for security.
In order to make beaker fips compliant, we flag this md5 use as not used
for security. Later we should invest trying to use uuids instead.
This change is based on this upstream PR:
https://github.com/bbangert/beaker/pull/241

Signed-off-by: Julien Doche <julien.doche@datadoghq.com>